downsize_encode

+++ xdg-user-dirs/media/bin/downsize_encode.sh
@@ -2,24 +2,6 @@
 source $(dirname $0)/00.conf
 export DISPLAY=:0
 
-cron_time=$(sed -e 's/:/ /g' <<< $MC_CRON_TIME)
-awk -v cron_time="$cron_time" 'BEGIN {
-    date = strftime("%Y %m %d")
-    date_time = sprintf("%s %s", date, cron_time)
-    cron_epoch = mktime(date_time)
-
-    if (60 * 60 * 3 < cron_epoch - systime()) {
-        exit 0
-    }
-    else {
-        exit 1
-    }
-}'
-
-if [ $? -eq 1 ];then
-    exit
-fi
-
 lock_file=/tmp/downsize_encode_lock
 
 lockfile-create $lock_file
@@ -51,6 +33,7 @@ if [ -n "$xml" ];then
     encode_width=$(xmlsel -t -m //encode-width   -v .   ${MC_DIR_ENCODING}/${job_file_xml})
     encode_height=$(xmlsel -t -m //encode-height -v .   ${MC_DIR_ENCODING}/${job_file_xml})
     encode_bitrate=$(xmlsel -t -m //encode-bitrate -v . ${MC_DIR_ENCODING}/${job_file_xml})
+    rec_time=$(xmlsel -t -m //rec-time -v .             ${MC_DIR_ENCODING}/${job_file_xml})
 
     log "d_encode start: $title $(hard_ware_info)"
 
@@ -62,14 +45,19 @@ if [ -n "$xml" ];then
 
     time_start=$(awk 'BEGIN { print systime() }')
 
-    nice -n 10 \
+    fifo_dir=/tmp/pt3/fifo
+    mkdir -p $fifo_dir
+    fifo_ffmpeg=${fifo_dir}/ffmpeg_$$
+    mkfifo -m 644 $fifo_ffmpeg
+
+    nice -n 5 \
     gst-launch-1.0 -q \
-     filesrc location=$input_ts_file \
+     filesrc location=$fifo_ffmpeg \
      ! video/mpegts \
      ! tsdemux name=demux \
      demux. \
             ! queue \
-              max-size-buffers=10000 \
+              max-size-buffers=5000 \
               max-size-time=0 \
               max-size-bytes=0 \
             ! mpegvideoparse \
@@ -79,6 +67,8 @@ if [ -n "$xml" ];then
               deinterlace-method=bob \
               scale-method=fast \
               height=$encode_height \
+            ! videorate \
+              max-rate=30 \
             ! vaapih264enc \
                tune=high-compression \
                rate-control=cqp \
@@ -87,25 +77,61 @@ if [ -n "$xml" ];then
             ! mux. \
      demux. \
             ! queue \
-              max-size-buffers=10000 \
+              max-size-buffers=5000 \
               max-size-time=0 \
               max-size-bytes=0 \
-            ! faad plc=true \
-            ! audioconvert \
-            ! 'audio/x-raw,channels=6' \
-            ! faac rate-control=ABR \
+            ! aacparse \
             ! mux. \
-     matroskamux name=mux min-index-interval=10000000000 ! filesink location=${MC_DIR_MP4}/${job_file_mkv}
+     matroskamux name=mux min-index-interval=10000000000 ! filesink location=${MC_DIR_MP4}/${job_file_mkv} &
+    pid_gst=$!
 
+    nice -n 5 \
+    ffmpeg -y -analyzeduration 2M \
+    -loglevel quiet \
+    -dual_mono_mode sub \
+    -i $input_ts_file \
+    -threads 1 \
+    -vcodec copy \
+    -acodec aac \
+    -f mpegts $fifo_ffmpeg &
+    pid_ffmpeg=$!
+
+
+    (
+        sleep $((rec_time * 2))
+
+        sleep 10
+        kill -TERM $pid_ffmpeg > /dev/null 2>&1
+        sleep 10
+        kill -KILL $pid_ffmpeg > /dev/null 2>&1
+
+        sleep 10
+        kill -TERM $pid_gst > /dev/null 2>&1
+        sleep 10
+        kill -KILL $pid_gst > /dev/null 2>&1
+    ) &
+
+    wait $pid_ffmpeg
+    wait $pid_gst
+
+    rm -f $fifo_ffmpeg
+
+    ffprobe -show_format ${MC_DIR_MP4}/${job_file_mkv} > /dev/null 2>&1
     if [ $? -eq 0 ];then
+        duration=$(ffprobe -show_format ${MC_DIR_MP4}/${job_file_mkv} 2> /dev/null | grep ^duration= | awk -F = '{ printf("%d\n", $2) }')
+        integrity=$(($rec_time - $duration))
+        if [ "$integrity" -lt 180 ];then
+            time_end=$(awk 'BEGIN { print systime() }')
+            (( took = (time_end - time_start) / 60 ))
 
-        time_end=$(awk 'BEGIN { print systime() }')
-        (( took = (time_end - time_start) / 60 ))
+            /bin/rm ${MC_DIR_ENCODING}/${job_file_xml}
+            stat --format=%s ${MC_DIR_MP4}/${job_file_mkv} > ${MC_DIR_FILE_SIZE}/${job_file_mkv}
 
-        /bin/rm ${MC_DIR_ENCODING}/${job_file_xml}
-        stat --format=%s ${MC_DIR_MP4}/${job_file_mkv} > ${MC_DIR_FILE_SIZE}/${job_file_mkv}
-
-        log "d_encode end: $took min $title $(hard_ware_info)"
+            log "d_encode end: $took min $title $(hard_ware_info)"
+        else
+            log "d_encode failed: $title $(hard_ware_info)"
+            /bin/mv ${MC_DIR_ENCODING}/${job_file_xml} $MC_DIR_FAILED
+        fi
     else
         log "d_encode failed: $title $(hard_ware_info)"
         /bin/mv ${MC_DIR_ENCODING}/${job_file_xml} $MC_DIR_FAILED
